train_model.py
""" import tensorflow as tf
import tensorflow_datasets as tfds
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Conv2D, MaxPooling2D, Flatten, Dense, Input

# Load the COCO dataset
dataset, info = tfds.load('coco/2017', with_info=True, split='train', data_dir='coco')

# Define a function to preprocess the images and labels
def preprocess(data):
    image = data['image']
    image = tf.image.resize(image, (150, 150))  # Resize to the desired size
    image = image / 255.0  # Normalize pixel values
    label = data['objects']['label']
    return image, label

# Apply the preprocessing function to the dataset
dataset = dataset.map(preprocess)

# Batch and shuffle the dataset
batch_size = 32
dataset = dataset.shuffle(buffer_size=1000).batch(batch_size).prefetch(tf.data.experimental.AUTOTUNE)

# Define a simple CNN model
model = Sequential([
    Input(shape=(150, 150, 3)),
    Conv2D(32, (3, 3), activation='relu'),
    MaxPooling2D((2, 2)),
    Conv2D(64, (3, 3), activation='relu'),
    MaxPooling2D((2, 2)),
    Conv2D(128, (3, 3), activation='relu'),
    MaxPooling2D((2, 2)),
    Flatten(),
    Dense(512, activation='relu'),
    Dense(info.features['objects']['label'].num_classes, activation='softmax')  # Adjust for multi-class classification
])

# Compile the model
model.compile(optimizer='adam',
              loss='sparse_categorical_crossentropy',
              metrics=['accuracy'])

# Train the model
model.fit(dataset, epochs=10)  # Adjust the number of epochs as needed

# Save the model
model.save('coco_trained_model.h5') """

import tensorflow as tf
from tensorflow.keras.models import load_model
from yolov4.tf import YOLOv4

# Define the path to the dataset
dataset_path = 'path/to/your/dataset'

# Load the dataset using TensorFlow Datasets or any other method
# For example, using TensorFlow Datasets:
import tensorflow_datasets as tfds
dataset, info = tfds.load('coco/2017', with_info=True, split='train', data_dir='coco')

# Define a function to preprocess the images and labels
def preprocess(data):
    image = data['image']
    image = tf.image.resize(image, (416, 416))  # Resize to the desired size for YOLO
    image = image / 255.0  # Normalize pixel values
    label = data['objects']['label']
    return image, label

# Apply the preprocessing function to the dataset
dataset = dataset.map(preprocess)

# Batch and shuffle the dataset
batch_size = 8
dataset = dataset.shuffle(buffer_size=1000).batch(batch_size).prefetch(tf.data.experimental.AUTOTUNE)

# Initialize YOLOv4 model
yolo = YOLOv4(tiny=True)

# Load pre-trained weights
yolo.classes = "coco.names"
yolo.input_size = (416, 416)
yolo.batch_size = batch_size
yolo.make_model()
yolo.load_weights("yolov4-tiny.weights", weights_type="yolo")

# Compile the model
yolo.model.compile(optimizer='adam',
                   loss='sparse_categorical_crossentropy',
                   metrics=['accuracy'])

# Train the model
yolo.model.fit(dataset, epochs=10)  # Adjust the number of epochs as needed

# Save the model
yolo.model.save('yolo_trained_model.h5')