requirements.txt
numpy
opencv-python
matplotlib
tensorflow
tensorflow_datasets
tensorflow-yolov4-tflite